+++ packages/custom_lint_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased patch
+
+- `matcherNormalizedPrioritizedSourceChangeSnapshot` now optionally allows specifying a `JsonEncoder`.
+
 ## 0.5.8 - 2024-01-09
 
 Added an optional `customPath` to the various `ChangeReporter` methods (thanks to @laurentschall)